refactor: LottoController를 controller 폴더안에 들어가도록 경로 변경
